MVC
Separated functions into ‘core’ and ‘controllers’. Set .htaccess to run
all urls off the index.php, requesting controllers and views accordingly
