+++ app/.htaccess
@@ -1,5 +1,7 @@
-RewriteEngine on
-RewriteCond %{REQUEST_FILENAME} !-f
+Options -MultiViews
+RewriteEngine On
+
 RewriteCond %{REQUEST_FILENAME} !-d
-RewriteCond %{REQUEST_FILENAME}.php -f
-RewriteRule !.*\.php$ %{REQUEST_FILENAME}.php [L]
+RewriteCond %{REQUEST_FILENAME} !-f
+
+RewriteRule ^(.+)$ index.php?url=$1 [QSA,L]
