+++ app/controllers/new-pet.php
@@ -1,0 +1,5 @@
+<?php require '../core/functions.php';
+
+submit_pet();
+
+?>
