--- app/functions/i-own-it.php
@@ -1,1 +1,0 @@
-<?php require 'functions.php'; ?>
