--- app/functions/new-found-pet.php
@@ -1,1 +1,0 @@
-<?php require 'functions.php'; ?>
