--- app/functions/new-pet.php
@@ -1,3 +1,0 @@
-<?php require 'functions.php'; 
-	submit_pet();
-?>
