+++ app/partials/header.php
@@ -1,4 +1,4 @@
-<?php include('functions/functions.php');?>
+<?php include('core/functions.php');?>
 <!doctype html>
 <html lang="en">
 <head>
