+++ app/views/found.php
@@ -1,8 +1,3 @@
-<?php
-    $page="found";
-    include("partials/header.php");
-?>
-
 <main class="l-main" role="main">
 
     <h2 class="heading heading--main h-spacing-base">Found</h2>
@@ -20,7 +15,3 @@
     </form>
 
 </main>
-
-<?php
-    include("partials/footer.php");
-?>
